Today's changes
git-svn-id: https://svn.code.sf.net/p/plexil/code/branches/plexil-x@4535 d2be67b0-814c-0410-949f-d215b6b29573

+++ src/value/CommandHandle.hh
@@ -35,7 +35,7 @@ namespace PLEXIL
   /**
    * @brief Command handle state enumeration.
    */
-  enum CommandHandleValue : uint16_t {
+  enum CommandHandleValue : uint8_t {
     NO_COMMAND_HANDLE = 48,
       COMMAND_SENT_TO_SYSTEM,
       COMMAND_ACCEPTED,
